Fixed types for tests (for print, for lenghts, etc)

+++ src/test_commons.h
@@ -7,7 +7,7 @@
 
 typedef enum {
   INT,
-  ULL,
+  UINT,
   DOUBLE,
   LDOUBLE,
   STRING,
